// packages/core-domain/src/enums/PolicyRuleType.ts
export enum PolicyRuleType {
  SIBLING_COHESION = 'SIBLING_COHESION',
  MIN_BLOCK_LENGTH = 'MIN_BLOCK_LENGTH',
  SCHOOL_NIGHT_ROUTINE = 'SCHOOL_NIGHT_ROUTINE',
  TRAVEL_DISTANCE_LIMIT = 'TRAVEL_DISTANCE_LIMIT',
  EXCHANGE_LOCATION = 'EXCHANGE_LOCATION',
  ACTIVITY_COMMITMENT = 'ACTIVITY_COMMITMENT',
}